norm complient (somewhat)
